server/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
python-dotenv
PyJWT
Werkzeug
Pillow
python-multipart 